chore: app.tsx navigator 컴포넌트로 분리
