+++ src/components/StackNavigator/index.tsx
@@ -1,0 +1,19 @@
+import React from "react";
+import { createStackNavigator } from "@react-navigation/stack";
+import { Home } from "@/pages/Home";
+
+const Stack = createStackNavigator();
+
+function HomeStack() {
+    return (
+        <Stack.Navigator>
+            <Stack.Screen
+                name="Home"
+                component={Home}
+                options={{ headerShown: false }}
+            />
+        </Stack.Navigator>
+    );
+}
+
+export default HomeStack;
